+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Upgrading to 3.6.x versions requires that you upgrade to latest 3.5.x version fi
 - Remove email parser overhead on Sphinx search (@vladsf, #487)
 - Make issue navigation respect user preference, default off (performance!) (@glensc, #485, #488)
 - Replace `pear/text_diff` with `sebastian/diff` (@glensc, #489)
+- Increase default height of Issue Reply pop-up (@vladsf, #470)
 
 [3.6.2]: https://github.com/eventum/eventum/compare/v3.6.1...master